Create RESTful server handling "people" data
This is the first commit in a project that 1. uploads a CSV file into a
MySQL database and opens up REST access to query said database.

Included in the README is more information about the application,
configuration and setup instructions, instructions on starting the app,
using the app, and additional things to improve the data model and
application.